Migrate to netstandard and xamarin forms pre
