+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,8 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\BottomBar.Droid",
+    "\\BottomBar.XamarinForms.NetStandard"
+  ],
+  "PreviewInSolutionExplorer": false
+}
