+++ BottomBar.Droid/packages.config
@@ -69,5 +69,5 @@
   <package id="Xamarin.Android.Support.v7.RecyclerView" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Vector.Drawable" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Build.Download" version="0.4.4" targetFramework="monoandroid71" />
-  <package id="Xamarin.Forms" version="2.3.5.239-pre3" targetFramework="monoandroid71" />
+  <package id="Xamarin.Forms" version="2.4.0.269-pre2" targetFramework="monoandroid71" />
 </packages>
