+++ BottomBarXFExample.NetStandard/App.xaml
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Application
+	xmlns="http://xamarin.com/schemas/2014/forms"
+	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+	x:Class="BottomBarXFExample.App">
+	<Application.Resources>
+	</Application.Resources>
+</Application>
