+++ BottomBarXFExample.NetStandard/TabPage.xaml
