+++ BottomBarXFExampleXaml.NetStandard/App.xaml
