+++ BottomBarXFExampleXaml.NetStandard/App.xaml.cs
