+++ BottomBarXFExampleXaml.NetStandard/BarPage.xaml
