+++ BottomBarXFExampleXaml.NetStandard/BarPage.xaml.cs
