+++ BottomBarXFExampleXaml.NetStandard/TabPage.xaml
