+++ BottomBarXFExampleXaml.NetStandard/TabPage.xaml.cs
