--- example/BottomBarXFExample/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Xamarin.Forms" version="2.3.5.239-pre3" targetFramework="portable45-net45+win8+wp8" />
-</packages>
